typingGmaeLogic切り分け
